backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<3.16
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
